pyproject.toml
[tool.poetry]
name = "cc"
version = "0.1.0"
description = ""
authors = ["emotzart <johnshuabratsk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
dotmap = "^1.3.30"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
